save location

+++ ITCMac/FeedItem.swift
@@ -5,6 +5,6 @@ struct FeedItem: Identifiable, Codable {
     var id: String = ""
     var title: String = ""
     var details: String = ""
-    var lat: Float? = nil
-    var long: Float? = nil
+    var lat: Double? = nil
+    var long: Double? = nil
 }
